refactor: move populate_demo_data into src/demo.rs (#494)
The ~555-line demo-data builder moves out of app.rs into a dedicated src/demo.rs
as a free function over &mut App; app.rs keeps a one-line delegation shim so
--demo mode (main.rs) and the UI snapshot tests (ui/mod.rs) call it unchanged.

This is a verbatim relocation (self -> app only). The UI snapshot tests render
this exact demo data at a fixed date, so the full insta suite passing confirms
the content is byte-for-byte preserved. app.rs drops ~553 lines.

Remaining for #494: splitting the inline test module out of app.rs.

No App fields changed; field-count ratchet unaffected.
